[IpkgInstaller] * move IpkgInstaller from Softwaremanager plugin to Screens/Opkg
